feat: playfab persistant data (WIP)

+++ Assets/Scripts/Enemy/EnemyHealth.cs
@@ -5,18 +5,18 @@ namespace Enemy
 {
     public class EnemyHealth : MonoBehaviour, IDamageable
     {
-        [SerializeField] private int maxHealth;
-        private int _currentHealth;
+        [SerializeField] private float maxHealth;
+        private float _currentHealth;
 
-        public event Action<int, int> InitEvent;
-        public event Action<int, int> OnDamagedEvent;
+        public event Action<float, float> InitEvent;
+        public event Action<float, float> OnDamagedEvent;
         public event Action OnDiedEvent;
 
         private void Awake() => _currentHealth = maxHealth;
 
         private void Start() => InitEvent?.Invoke(_currentHealth, maxHealth);
 
-        public void TakeDamage(int amount)
+        public void TakeDamage(float amount)
         {
             _currentHealth -= amount;
