+++ Assets/Scripts/PlayerWeapons/Bullet.cs
@@ -5,10 +5,10 @@ namespace PlayerWeapons
     [RequireComponent(typeof(Rigidbody))]
     public class Bullet : MonoBehaviour
     {
-        private int _damage;
+        private float _damage;
         private Rigidbody _rigidbody;
 
-        public void SetDamage(int damage) => _damage = damage;
+        public void SetDamage(float damage) => _damage = damage;
 
         public void AddForce(Vector3 dir, int speed) => _rigidbody.AddForce(dir * speed);
